feat: add `insert_if_new` (#14397)
Often there are reasons to insert some components (e.g. Transform)
separately from the rest of a bundle (e.g. PbrBundle).
However `insert` overwrites existing components, making this difficult.

This PR adds the method `insert_if_new` to EntityMut and Commands, which
is the same as `insert` except that the old component is kept in case of
conflicts.

It also renames some internal enums (from `ComponentStatus::Mutated` to
`Existing`), to reflect the possible change in meaning.

- Did you test these changes? If so, how?

Added basic unit tests; used the new behavior in my project.

- Are there any parts that need more testing?

There should be a test that the change time isn't set if a component is
not overwritten; I wasn't sure how to write a test for that case.